// src/app/models/requests/monitoring-rule-condition.request.model.ts
export class MonitoringRuleConditionRequestModel {
  comparison_operator: string;
  created_at: string;
  id: number;
  list_item_value: string;
  logical_operator: string;
  monetary_value: number;
  number_per_hour: number;
  numeric_without_decimal_places_value: number;
  updated_at: string;
  value: string;
  variable_name: string;
}