// src/app/models/response/monitoring-rule-condition.response.model.ts
export class MonitoringRuleConditionResponseModel {
  comparison_operator: string;
  created_at: string;
  id: number;
  list_item_value: string;
  logical_operator: string;
  monetary_value: number;
  number_per_hour: number;
  numeric_without_decimal_places_value: number;
  updated_at: string;
  value: string;
  variable_name: string;
}